+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+- Replaced slow tooltip placement workaround with a more efficient workaround
+
 # Version 0.5.0
 
 - **Updated to Bevy 0.17**